Refactor Parameters class to use BigInteger for q and alpha

+++ src/Parameters.java
@@ -1,23 +1,24 @@
+import java.math.BigInteger;
 
 public class Parameters {
  
     // Private fields for q (large prime number) and alpha (primitive root)
-    private int q;
-    private int alpha;
+    private BigInteger q;
+    private BigInteger alpha;
 
     // Constructor that sets q and alpha
-    public Parameters(int q, int alpha) {
+    public Parameters(BigInteger q, BigInteger alpha) {
        this.q = q;       
         this.alpha = alpha;
     }
 
     // Getter for q
-    public int getQ() {
+    public BigInteger getQ() {
         return q;
     }
 
     // Getter for alpha
-    public int getAlpha() {
+    public BigInteger getAlpha() {
         return alpha;
     }
